V4L/DVB (10138): v4l2-ioctl: change to long return type to match unlocked_ioctl.
Since internal to v4l2 the ioctl prototype is the same regardless of it
being called through .ioctl or .unlocked_ioctl, we need to convert it all
to the long return type of unlocked_ioctl.

Thanks to Jean-Francois Moine for posting an initial patch for this and
thus bringing it to our attention.

Cc: Jean-Francois Moine <moinejf@free.fr>

+++ drivers/media/common/saa7146_fops.c
@@ -309,7 +309,7 @@ static int fops_release(struct file *file)
 	return 0;
 }
 
-static int fops_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
+static long fops_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 {
 /*
 	DEB_EE(("file:%p, cmd:%d, arg:%li\n", file, cmd, arg));
